Added /queen leaderboard

+++ src/main/java/com/zerek/featherqueenofthehill/listeners/EntityDamageByEntityListener.java
@@ -2,7 +2,6 @@ package com.zerek.featherqueenofthehill.listeners;
 
 import com.zerek.featherqueenofthehill.FeatherQueenOfTheHill;
 import com.zerek.featherqueenofthehill.managers.QueenManager;
-import org.bukkit.entity.HumanEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -22,10 +21,10 @@ public class EntityDamageByEntityListener implements Listener {
     public void onEntityDamageByEntity(EntityDamageByEntityEvent event){
         if (qm.isQueenOnline() && qm.getQueen().getPlayer().hasPermission("feather.queenofthehill.atspawn")){
             if (event.getDamager() == qm.getQueen() && event.getEntity() instanceof Player){
-                qm.addScore(event.getDamage()/75.0);
+                qm.addScore((float) (event.getDamage()/75.0));
             }
             if (event.getEntity() == qm.getQueen() && event.getDamager() instanceof Player){
-                qm.addScore(event.getDamage()/100.0);
+                qm.addScore((float) (event.getDamage()/100.0));
             }
         }
     }
